chore: comment from peer-review.
logion-network/logion-internal#480
